# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Messenger API
  description: Self-hosted messaging server for humans to interact with AI agents
  version: 0.2.0
  contact:
    name: Joel Claw
    url: https://joelclaw.lu
servers:
  - url: http://localhost:8080
    description: Local development
security: []
paths:
...
  /conversations/delete:
    post:
      summary: Delete a conversation
      operationId: deleteConversation
      tags: [Conversations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversation_id]
              properties:
                conversation_id:
                  type: string
      responses:
        '200':
          description: Conversation deleted
        '401':
          description: Unauthorized (not owner)
        '404':
          description: Conversation not found
...
  /messages/search:
    get:
      summary: Search messages
      operationId: searchMessages
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        message_id:
          type: string
        conversation_id:
          type: string
        sender_type:
          type: string
          enum: [user, agent]
        sender_id:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
          nullable: true
        edited:
          type: boolean
        deleted:
          type: boolean
...